integrate and fix bugs
