+++ src/main/java/funny/gobang/model/Point.java
@@ -49,4 +49,13 @@ public class Point {
         result = 31 * result + y;
         return result;
     }
+
+    @Override
+    public String toString() {
+        final StringBuilder sb = new StringBuilder("Point{");
+        sb.append("x=").append(x);
+        sb.append(", y=").append(y);
+        sb.append('}');
+        return sb.toString();
+    }
 }
