+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-search.depth=6
+search.depth=2
